pyproject.toml
[project]
name = "senior-ai-engineer-test"
version = "0.1.0"
description = "Software for monitoring lab environment"
authors = [
    {name = "Matt Freeland",email = "matthew_freeland@yahoo.co.uk"}
]
readme = "README.md"
requires-python = "^3.11.9"
dependencies = [
    "pylint (>=3.3.7,<4.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pdoc3 (>=0.11.6,<0.12.0)",
    "pylint-pydantic (>=0.3.5,<0.4.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "numpydantic (>=1.6.9,<2.0.0)"
]

[tool.poetry]
packages = [
    { include = "lab_monitor", from = "src" }
]


[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
pylint = "^3.3.7"
pdoc3 = "^0.11.6"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=lab_monitor --cov-report=term --cov-report=html"
testpaths = ["tests"]

[tool.flake8]
max-line-length = 120
exclude = [".cache", ".poetry", "venv"]

[tool.pylint.MASTER]
disable = ["E0401", "W0613", "R0902", "R0912", "R0913", "R0914", "W0511", "C0204", "R1702", "W0703", "W0107", "R0903"]
exclude = [".cache", ".poetry", "venv"]

[tool.pylint.FORMAT]
max-line-length = 120

[tool.coverage.run]
source = "lab_monitor"
